# services/api/cmd/server/docs/swagger.yaml
basePath: /
definitions:
  github_com_gianghp123_SonaVoice_api_internal_core_enums.MessageRole:
    enum:
    - user
    - assistant
    type: string
    x-enum-varnames:
    - MessageRoleUser
    - MessageRoleAssistant
...
  github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError'
      meta:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.Meta'
      success:
        type: boolean
    type: object
...
  ? github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-array_github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_res_MessageRes
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_res.MessageRes'
        type: array
      error:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError'
      meta:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.Meta'
      success:
        type: boolean
    type: object
...
  ? github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_SessionConfigRes
  : properties:
      data:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.SessionConfigRes'
      error:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_errors.AppError'
      meta:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.Meta'
      success:
        type: boolean
    type: object
...
  github_com_gianghp123_SonaVoice_api_internal_core_response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_req.CreateMessagesReq:
    properties:
      messages:
        items:
          $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_req.MessageItem'
        type: array
    required:
    - messages
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_req.MessageItem:
    properties:
      role:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_enums.MessageRole'
      transcript:
        type: string
      was_interrupted:
        type: boolean
    required:
    - role
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_res.MessageRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_enums.MessageRole'
      session_id:
        type: string
      transcript:
        type: string
      was_interrupted:
        type: boolean
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.ConfigPayload:
    properties:
      enabled:
        type: boolean
      limits:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.LimitsConfig'
      reset_timezone:
        type: string
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.LimitsConfig:
    properties:
      session:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.SessionLimitConfig'
      user:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.UserLimitConfig'
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.SessionLimitConfig:
    properties:
      max_session_lockTTL:
        type: integer
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.UserLimitConfig:
    properties:
      daily_request_count:
        type: integer
      daily_voice_seconds:
        type: integer
    type: object
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_req.SessionConfigReq:
    properties:
      config:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.ConfigPayload'
    type: object
...
  github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res.SessionConfigRes:
    properties:
      config:
        $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos.ConfigPayload'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Session API
  version: "1.0"
paths:
...
  /sessions/{sessionId}/messages:
    get:
      consumes:
      - application/json
      description: Get paginated list of messages for a session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-array_github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_res_MessageRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: List messages in a session
      tags:
      - message
    post:
      consumes:
      - application/json
      description: Create new messages for a session (internal endpoint)
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Messages to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_req.CreateMessagesReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-array_github_com_gianghp123_SonaVoice_api_internal_modules_message_dtos_res_MessageRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      summary: Create messages in a session
      tags:
      - message
...
  /sessions/config:
    get:
      description: Retrieve the session configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_SessionConfigRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      summary: Get session config
      tags:
      - session-config
    put:
      consumes:
      - application/json
      description: Update the session configuration (admin only)
      parameters:
      - description: Session config payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_req.SessionConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-github_com_gianghp123_SonaVoice_api_internal_modules_session_dtos_res_SessionConfigRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_gianghp123_SonaVoice_api_internal_core_response.BaseResponse-any'
      security:
      - BearerAuth: []
      summary: Update session config
      tags:
      - session-config
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"